Fixed after refactoring
