+++ app/views/admin/form_answers/company_details/_this_entry_relates_to_form.html.slim
@@ -1,8 +1,4 @@
-- account_name = "This entry relates to"
-- trade_ops = ["products", "services"]
-- innovation_ops = ["product", "service", "business_model"]
-- development_ops = ["product", "service", "management_approach"]
-- mobility_ops = ["mentoring", "career_opportunities_accessibility", "workplace_fostering"]
+- title = "This entry relates to"
 
 - unless @form_answer.promotion?
   - if user_can_edit(@form_answer, :this_entry_relates_to)
@@ -13,42 +9,19 @@
         = hidden_field_tag :section, "this_entry_relates_to"
 
         .form-container
-          label.form-label = account_name
+          label.form-label = title
 
           .form-value
             p
               - if @form_answer.this_entry_relates_to.present?
                 ul
-                  - AdminThisEntryRelatesToOptionsHelper::DATA.each do |item|
-                    - @form_answer.this_entry_relates_to.each do |value|
-                      - if value == item.keys.join(" ")
-                        li
-                          = item.values.join(" ")
+                  = render "admin/form_answers/company_details/entry_relates_to_value"
               - else
                 em.text-muted No "this entry relates to" added.
 
           .form-fields.form-block
-            - if @form_answer.innovation? && options_available(innovation_ops) || @form_answer.trade? && options_available(trade_ops)
-              - values = @form_answer.document["application_relate_to_header"]
-
-              - @check_options.each_with_index do |check_option, placement|
-                - value_selected = values.is_a?(Array) && values.detect { |el| el['type'].to_s == check_option[0].to_s }
-
-                .form-control
-                  label
-                    input type="checkbox" name="form_answer[data_attributes][application_relate_to_header][#{placement}][type]" value="#{check_option[0]}" checked=('checked' if value_selected)
-                    |  #{check_option[1]}
-
-            - if @form_answer.development? && options_available(development_ops) || @form_answer.mobility? && options_available(mobility_ops)
-              - values = @form_answer.document["application_relate_to"]
-
-              - @check_options.each_with_index do |check_option, placement|
-                - value_selected = values.is_a?(Array) && values.detect { |el| el['type'].to_s == check_option[0].to_s }
-
-                .form-control
-                  label
-                    input type="checkbox" name="form_answer[data_attributes][application_relate_to_header][#{placement}][type]" value="#{check_option[0]}" checked=('checked' if value_selected)
-                    |  #{check_option[1]}
+            = render "admin/form_answers/company_details/entry_relates_to_checkboxes", awards: ["innovation", "trade"]
+            = render "admin/form_answers/company_details/entry_relates_to_checkboxes", awards: ["development", "mobility"]
 
             .form-actions.text-right
               = link_to "Cancel", "#", class: "btn btn-default form-cancel-link if-no-js-hide"
@@ -62,11 +35,7 @@
   - else
     .form-group
       .form-container
-        label.form-label = account_name
+        label.form-label = title
         p
           ul
-            - AdminThisEntryRelatesToOptionsHelper::DATA.each do |item|
-              - @form_answer.this_entry_relates_to.each do |value|
-                - if value == item.keys.join(" ")
-                  li
-                    = item.values.join(" ")
+            = render "admin/form_answers/company_details/entry_relates_to_value"
